+++ src/network.h
@@ -20,8 +20,8 @@ private:
     const TActivation neuron_activation_type;
     const TInteraction neuron_interaction_type;
 
-    std::vector<double> biases;
-    std::vector<std::vector<double> > weights;
+    std::vector<float> biases;
+    std::vector<std::vector<float> > weights;
     std::vector<std::vector<int64_t> > connected_neuron_ids;
     std::vector<Neuron> neurons;
     int64_t tauref;
@@ -32,11 +32,11 @@ private:
 
     void generate_connected_neuron_ids();
     std::vector<int64_t> get_update_inds();
-    double get_potential_for_neuronid(int64_t id);
+    float get_potential_for_neuronid(int64_t id);
 
 public:
-    Network(std::vector<double> &_biases,
-            std::vector<std::vector<double> > &_weights,
+    Network(std::vector<float> &_biases,
+            std::vector<std::vector<float> > &_weights,
             std::vector<int64_t> &_initialstate,
             Config config);
     ~Network() {};
@@ -45,12 +45,12 @@ public:
 
     // std::vector<bool> get_state();
     void produce_header(std::ostream& stream);
-    void produce_output(std::ostream& stream, double T, double Iext);
+    void produce_output(std::ostream& stream, float T, float Iext);
     void produce_summary(std::ostream& stream);
     void get_state();
     void get_internalstate();
-    void update_state(double T);
-    void update_state(double T, double Iext);
+    void update_state(float T);
+    void update_state(float T, float Iext);
     bool _check_consistency();
 };
